Fix bugs and refactor design pattern
