+++ app.rb
@@ -2,18 +2,22 @@ require_relative 'book'
 require_relative 'label'
 require_relative 'item'
 require_relative 'game'
-require_relative './store/preserve_data'
 require_relative 'author'
 require 'json'
 require 'date'
-require_relative './src/genre'
-require_relative './src/music_album'
-require_relative './handler'
+require_relative 'genre'
+require_relative 'music_album'
+# Documentation for the App class
+require_relative 'book'
+require_relative 'label'
+
 class App
-  attr_accessor :books
+  attr_accessor :books, :authors, :games
 
   def initialize
     @books = []
+    @authors = []
+    @games = []
     @labels = [Label.new('Love at the end of the day.', 'Blue'), Label.new('The love of my lives.', 'Grey')]
     base = "#{Dir.pwd}/data"
     books_reader = File.read("#{base}/books.json")
@@ -24,56 +28,56 @@ class App
     end
   end
 
+  # create a list of BOOK
   def list_books
+    if @books.empty?
+      puts 'Oops! the library is empty, please add books'
+      return
+    end
     @books.each do |book|
-      puts "#{book['id']} - #{book['author']} - #{book['published_date']}"
+      puts "publisher: #{book.publisher}, cover_state: #{book.cover_state}, published_date: #{book.published_date}"
     end
   end
 
-  def list_labels
-    @labels.each do |label|
-      puts "#{label['id']} - #{label['title']} - #{label['color']}"
-    end
-  end
-
+  # create a book
   def create_book
-    puts 'Enter author'
-    author = gets.chomp
-    puts 'Enter published date (YYYY-MM-DD))'
-    published_date = gets.chomp
-    puts 'Enter publisher'
+    print 'Publisher: '
     publisher = gets.chomp
-    puts 'Enter cover state'
+
+    print 'Cover State: '
     cover_state = gets.chomp
-    puts 'Enter label title'
-    label_title = gets.chomp
-    puts 'Enter label color'
-    label_color = gets.chomp
-    label = create_label(label_title, label_color)
-    book = create_book_object(author, published_date, publisher, cover_state, label)
-    @books << book
-    @labels << label
-    @store.save_data(@books, './store/books.json')
-    @store.save_data(@labels, './store/labels.json')
+
+    print 'Published Date: '
+    published_date = gets.chomp
+
+    @books << Book.new(publisher, cover_state, published_date)
+
     puts 'Book created successfully'
   end
 
-  def create_label(title, color)
-    label = Label.new(title, color)
-    { 'id' => label.id, 'title' => label.title, 'color' => label.color }
+  # create a list of labels
+  def list_labels
+    if @labels.empty?
+      puts 'Oops! the library is empty, please add books'
+      return
+    end
+
+    @labels.each do |label|
+      puts "title: #{label.title}, color: #{label.color}"
+    end
   end
 
-  def create_book_object(author, published_date, publisher, cover_state, label)
-    book = Book.new(author, published_date, publisher, cover_state, label)
-    {
-      'id' => book.id,
-      'author' => book.author,
-      'published_date' => book.published_date,
-      'publisher' => book.publisher,
-      'cover_state' => book.cover_state,
-      'label' => book.label
-    }
-  end
+  # def create_book_object(author, published_date, publisher, cover_state, label)
+  #   book = Book.new(author, published_date, publisher, cover_state, label)
+  #   {
+  #     'id' => book.id,
+  #     'author' => book.author,
+  #     'published_date' => book.published_date,
+  #     'publisher' => book.publisher,
+  #     'cover_state' => book.cover_state,
+  #     'label' => book.label
+  #   }
+  # end
 
   def list_games
     if @games.empty?
@@ -121,10 +125,10 @@ class App
     last_name = gets.chomp
     author = create_author(first_name, last_name)
     authors << author
-    @store.save_data(authors, './store/authors.json')
+    @store.save_data(authors, './data/authors.json')
     game = create_game_object(game_name, last_played_at, publish_date, multiplayer)
     @games << game
-    @store.save_data(@games, './store/games.json')
+    @store.save_data(@games, './data/games.json')
     puts 'Game is created successfully'
   end
