--- store/authors.json
@@ -1,17 +1,0 @@
-[
-  {
-    "id": 180,
-    "first_name": "bvcx",
-    "last_name": "nbcc"
-  },
-  {
-    "id": 631,
-    "first_name": "Brhanu",
-    "last_name": "Hailu"
-  },
-  {
-    "id": 483,
-    "first_name": "marki",
-    "last_name": "msess"
-  }
-]
