--- store/books.json
@@ -1,86 +1,0 @@
-[
-  {
-    "id": 535,
-    "author": "author1",
-    "published_date": "2010-09-09",
-    "publisher": "pub",
-    "cover_state": "good",
-    "label": {
-      "id": 476,
-      "title": "fiction",
-      "color": "green"
-    }
-  },
-  {
-    "id": 256,
-    "author": "author2",
-    "published_date": "1998-07-30",
-    "publisher": "pubmed",
-    "cover_state": "bad",
-    "label": {
-      "id": 84,
-      "title": "title2",
-      "color": "orange"
-    }
-  },
-  {
-    "id": 345,
-    "author": "author3",
-    "published_date": "1765-03-14",
-    "publisher": "pubmed",
-    "cover_state": "bad",
-    "label": {
-      "id": 170,
-      "title": "science",
-      "color": "red"
-    }
-  },
-  {
-    "id": 858,
-    "author": "a",
-    "published_date": "2017-08-04",
-    "publisher": "pub",
-    "cover_state": "good",
-    "label": {
-      "id": 181,
-      "title": "history",
-      "color": "violet"
-    }
-  },
-  {
-    "id": 844,
-    "author": "yod",
-    "published_date": "220-09-25",
-    "publisher": "Biomed",
-    "cover_state": "good",
-    "label": {
-      "id": 157,
-      "title": "Science",
-      "color": "yellow"
-    }
-  },
-  {
-    "id": 973,
-    "author": "John",
-    "published_date": "1990-12-12",
-    "publisher": "Nasa",
-    "cover_state": "bad",
-    "label": {
-      "id": 918,
-      "title": "Robotic",
-      "color": "dark"
-    }
-  },
-  {
-    "id": 206,
-    "author": "a",
-    "published_date": "1888-02-02",
-    "publisher": "pub",
-    "cover_state": "bad",
-    "label": {
-      "id": 214,
-      "title": "iction",
-      "color": "red"
-    }
-  }
-]
