--- store/games.json
@@ -1,23 +1,0 @@
-[
-  {
-    "id": 146,
-    "game_name": "jhkjkk",
-    "last_played_at": "2023-5-2",
-    "publish_date": "2022-05-03",
-    "multiplayer": "jhhjhhj"
-  },
-  {
-    "id": 395,
-    "game_name": "valleyball",
-    "last_played_at": "2023-6-2",
-    "publish_date": "2022-05-03",
-    "multiplayer": "vlc"
-  },
-  {
-    "id": 601,
-    "game_name": "hwrhawi",
-    "last_played_at": "1-2-2023",
-    "publish_date": "2023-06-02",
-    "multiplayer": "mcccccc"
-  }
-]
