--- store/genres.json
@@ -1,10 +1,0 @@
-[
-    {
-      "genre_id" : 77,
-      "genre_name" : "yyyy"
-    },
-    {
-      "genre_id" : 570,
-      "genre_name" : "tttt"
-    }
-  ]
