--- store/labels.json
@@ -1,37 +1,0 @@
-[
-  {
-    "id": 476,
-    "title": "fiction",
-    "color": "green"
-  },
-  {
-    "id": 84,
-    "title": "title2",
-    "color": "orange"
-  },
-  {
-    "id": 170,
-    "title": "science",
-    "color": "red"
-  },
-  {
-    "id": 181,
-    "title": "history",
-    "color": "violet"
-  },
-  {
-    "id": 157,
-    "title": "Science",
-    "color": "yellow"
-  },
-  {
-    "id": 918,
-    "title": "Robotic",
-    "color": "dark"
-  },
-  {
-    "id": 214,
-    "title": "iction",
-    "color": "red"
-  }
-]
