--- store/music_albums.json
@@ -1,12 +1,0 @@
-[
-    {
-      "music_id" : 45,
-      "music_name" : "www",
-      "music_on_spotify" : true
-    },
-    {
-      "music_id" : 53,
-      "music_name" : "8",
-      "music_on_spotify" : true
-    }
-  ]
